requirements.txt
attr==0.3.2
ConfigParser==7.0.0
cryptography==43.0.0
Cython==3.0.10
Django==5.0.7
djangorestframework==3.15.2
dl==0.1.0
docutils==0.21.2
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.3
Jinja2==3.1.4
keyring==25.2.1
langchain==0.2.11
langchain_community==0.2.10
langchain_core==0.2.25
langchain_openai==0.1.19
lockfile==0.12.2
pdfminer.six
Pillow==10.4.0
pillow_heif==0.10.0
protobuf==5.27.2
pyOpenSSL==24.2.1
pywin32==306; sys_platform == 'win32'
railroad==0.5.0
Sphinx==8.0.2
thread==2.0.5
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
trove_classifiers==2024.7.2
truststore==0.9.1
unstructured==0.15.0
